+++ tests/Integration/Framework/Testing/ModuleDoubleFixture/Greeting/Domain/Greeter.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Integration\Framework\Testing\ModuleDoubleFixture\Greeting\Domain;
+
+final class Greeter
+{
+    public function __construct(
+        private readonly string $greeting,
+    ) {
+    }
+
+    public function greet(): string
+    {
+        return $this->greeting;
+    }
+}